+++ src/components/admin/Pagination.jsx
@@ -12,24 +12,22 @@ export default function Pagination({ currentPage, setCurrentPage, pageAmt }) {
     <nav>
       <ul css={ctn}>
         {pages.map((page, idx) => (
-          <>
-            <li key={page} style={{ margin: "0 5px" }}>
-              <button
-                onClick={() => handlePageClick(page)}
-                css={{
-                  fontSize: "20px",
-                  fontWeight: "700",
-                  lineHeight: "26px",
-                  textAlign: "center",
-                  color: page === currentPage ? "#1676BC" : "#999999",
-                  backgroundColor: "transparent",
-                }}
-              >
-                {page}
-              </button>
-            </li>
+          <li key={page + idx} css={li}>
+            <button
+              onClick={() => handlePageClick(page)}
+              css={{
+                fontSize: "20px",
+                fontWeight: "700",
+                lineHeight: "26px",
+                textAlign: "center",
+                color: page === currentPage ? "#1676BC" : "#999999",
+                backgroundColor: "transparent",
+              }}
+            >
+              {page}
+            </button>
             {pages.length - 1 > idx && <span css={divider}></span>}
-          </>
+          </li>
         ))}
       </ul>
     </nav>
@@ -43,8 +41,16 @@ const ctn = css`
   gap: 14px;
 `;
 
+const li = css`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 14px;
+`;
+
 const divider = css`
   width: 2px;
+  display: block;
   height: 18px;
   background: #d9d9d9;
 `;
